requirements.txt
customtkinter>=5.2.0
Pillow>=10.0.0
pystray>=0.19.0
mido>=1.3.0
python-rtmidi>=1.5.0
pyserial>=3.5
